Tests/LibWeb/Text/expected/css/getComputedStyle-url-var-serialization.txt

style.getPropertyValue("--test-url") = url('https://ladybird.org/')
